+++ samples/XCT.Sample/Pages/TestCases/Issue1883.xaml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<pages:BasePage xmlns="http://xamarin.com/schemas/2014/forms"
+                xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
+                xmlns:pages="clr-namespace:Xamarin.CommunityToolkit.Sample.Pages"
+                x:Class="Xamarin.CommunityToolkit.Sample.Pages.Views.Issue1883Page">
+    <StackLayout Spacing="10" Margin="20">
+        <Button Clicked="Issue1883" Text="Show snackbar"/>
+        <Label x:Name="StatusText"/>
+    </StackLayout>
+</pages:BasePage>
