+++ samples/XCT.Sample/Pages/Views/SnackBarPage.xaml.cs
@@ -151,36 +151,6 @@ namespace Xamarin.CommunityToolkit.Sample.Pages.Views
 			StatusText.Text = result ? "SnackBar is closed by user" : "SnackBar is closed by timeout";
 		}
 
-		async void Issue1883(object sender, EventArgs args)
-		{
-			var options = new SnackBarOptions
-			{
-				MessageOptions = new MessageOptions
-				{
-					Foreground = Color.DodgerBlue,
-					Padding = new Thickness(10, 20, 30, 40),
-					Message = "Pour supprimer un élément de votre liste hors connexion, faites glisser l'élément vers la droite"
-				},
-				BackgroundColor = Color.Aquamarine,
-				Duration = TimeSpan.FromSeconds(5),
-				CornerRadius = new Thickness(10, 20, 30, 40),
-				Actions = new[]
-				{
-					new SnackBarActionOptions
-					{
-						Text = "I've completed this",
-						BackgroundColor = (Color.Aqua),
-						ForegroundColor = Color.Brown,
-						Padding = new Thickness(40, 30, 20, 10),
-						Action = () => Task.FromResult(true)
-					},
-				}
-			};
-
-			var result = await this.DisplaySnackBarAsync(options);
-			StatusText.Text = result ? "SnackBar is closed by user" : "SnackBar is closed by timeout";
-		}
-
 		string GenerateLongText(int stringDuplicationTimes)
 		{
 			const string message = "It is a very long message to test multiple strings. A B C D E F G H I I J K L M N O P Q R S T U V W X Y Z";
